Implemented User profile updation by default

+++ src/main/java/org/jsp/cda/repository/FacultyProfileRepository.java
@@ -2,7 +2,8 @@ package org.jsp.cda.repository;
 
 import org.jsp.cda.entity.FacultyProfile;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.web.multipart.MultipartFile;
 
-public interface FacultyProfileRepository extends JpaRepository<FacultyProfile, Integer> {
+public interface FacultyProfileRepository extends JpaRepository<FacultyProfile, Long> {
 
 }
